--- packages/mask/src/plugins/Wallet/constants.ts
@@ -1,1 +1,0 @@
-export { PLUGIN_ID, UPDATE_CHAIN_STATE_DELAY } from '@masknet/plugin-wallet'
