+++ packages/web3-shared/base/src/helpers/constant.ts
