+++ packages/web3-shared/base/src/helpers/formatBalance.ts
