+++ packages/web3-shared/base/src/helpers/formatCurrency.ts
