+++ packages/web3-shared/base/src/helpers/formatElapsed.ts
