+++ packages/web3-shared/base/src/helpers/formatMarketCap.ts
