+++ packages/web3-shared/base/src/helpers/formatPercentage.ts
