+++ packages/web3-shared/base/src/helpers/formatSupply.ts
