+++ packages/web3-shared/base/src/index.ts
@@ -1,2 +1,2 @@
 export * from './specs/index.js'
-export * from './utils/index.js'
+export * from './helpers/index.js'
