+++ packages/web3-shared/solana/src/index.ts
@@ -2,4 +2,4 @@ export * from './sdk/index.js'
 export * from './types.js'
 export * from './hooks/index.js'
 export * from './constants/index.js'
-export * from './utils/index.js'
+export * from './helpers/index.js'
